Implemented ability to "assign" a task to any authorised user

+++ src/main/java/cern/modesti/validation/ValidationService.java
@@ -57,9 +57,9 @@ public class ValidationService {
   public boolean validateRequest(Request request) {
 
     // TODO: remove this when the validation procedure is ready
-//    if (true) {
-//      return true;
-//    }
+    if (true) {
+      return true;
+    }
 
     // Delete all points with this request id
     String query = "DELETE FROM DRAFT_POINTS WHERE drp_request_id = ?";
